email started

+++ public/games.php
@@ -2,6 +2,17 @@
 $customer_games = games::get($_GET['id']);
 $customer = customers::get($_GET['id']);
 $games = games::getAll();
+
+if ($_POST) {
+    games::add($_GET['id'], $_POST['game']);
+    header('location: games/?id=' . $_GET['id']);
+}
+
+if (isset($_GET['delete'])) {
+    games::delete('id', $_GET['delete']);
+    header('location: games/?id=' . $_GET['id']);
+
+}
 ?>
 
 <!DOCTYPE html>
@@ -10,13 +21,15 @@ $games = games::getAll();
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Ratjetoe - Customers</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/css/bootstrap.min.css" integrity="sha384-xOolHFLEh07PJGoPkLv1IbcEPTNtaed2xpHsD9ESMhqIYd0nLMwNLD69Npy4HI+N" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/css/bootstrap.min.css"
+          integrity="sha384-xOolHFLEh07PJGoPkLv1IbcEPTNtaed2xpHsD9ESMhqIYd0nLMwNLD69Npy4HI+N" crossorigin="anonymous">
 </head>
 <body>
 
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
     <a class="navbar-brand" href="customer">Ratjetoe</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
     </button>
 
@@ -44,9 +57,9 @@ $games = games::getAll();
             <select class="form-control" name="game" required>
                 <?php
 
-                foreach ($games as $game){
+                foreach ($games as $game) {
 
-                    echo"<option value='". $game['id'] ."'>". $game['name'] ." (" . $game['platform'] . ")</option>";
+                    echo "<option value='" . $game['id'] . "'>" . $game['name'] . " (" . $game['platform'] . ")</option>";
                 }
                 ?>
             </select>
@@ -64,12 +77,12 @@ $games = games::getAll();
             </thead>
             <tbody>
             <?php
-            foreach ($customer_games as $customer_game){
+            foreach ($customer_games as $customer_game) {
                 echo "
                 <tr>
-                    <td>". $customer_game['name'] ."</td>
-                    <td>". $customer_game['platform'] ."</td>
-                    <td><a class='btn btn-danger' href='?delete=" . $customer_game['id'] ."'>X</a></td>
+                    <td>" . $customer_game['name'] . "</td>
+                    <td>" . $customer_game['platform'] . "</td>
+                    <td><a class='btn btn-danger' onclick='return confirm(\"Are you sure?\")' href='?delete=" . $customer_game['id'] . "&id=" . $_GET['id'] . "'>X</a></td>
                 </tr>
                 ";
             }
@@ -82,8 +95,12 @@ $games = games::getAll();
 </div>
 
 
-
 <!-- Add Bootstrap JS and Popper.js -->
-<script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-Fy6S3B9q64WdZWQUiU+q4/2Lc9npb8tCaSX9FK7E8HnRr0Jz8D6OP9dO5Vg3Q9ct" crossorigin="anonymous"></script></body>
+<script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js"
+        integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj"
+        crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-Fy6S3B9q64WdZWQUiU+q4/2Lc9npb8tCaSX9FK7E8HnRr0Jz8D6OP9dO5Vg3Q9ct"
+        crossorigin="anonymous"></script>
+</body>
 </html>
